// src/common/enums/error-code.enum.ts
export enum AUTHENTICATION_EXIT_CODE {
  NO_TOKEN = 1001,
  INVALID_TOKEN = 1002,
  EXPIRED_TOKEN = 1003,
  USER_NOT_FOUND = 1004,
  PASSWORD_NOT_CORRECT = 1005,
}

export enum AUTHORIZATION_EXIT_CODE {
  ACCESS_DENIED = 2001,
}

export enum VALIDATION_EXIT_CODE {
  EMPTY = 3001,
  INVALID_FORMAT = 3002,
  UNIQUE_VALUE = 3003,
  INVALID_LENGTH = 3004,
  INVALID_CHARACTOR = 3005,
  NOT_FOUND = 3006,
  NO_MATCHING = 3007,
  INVALID_VALUE = 3008,
}

export enum DATABASE_EXIT_CODE {
  CANNOT_CONNECT = 4001,
  UNAUTHORIZE = 4002,
  TIMEOUT = 4003,
  OPERATOR_ERROR = 4004,
  LOST_CONNECT = 4005,
  UNIQUE_FIELD_VALUE = 4006,
  UNKNOW_VALUE = 4007,
  NO_CONTENT = 4008,
}

export enum FILE_EXIT_CODE {
  INVALID_EXTENSION = 5001,
  INVALID_SIZE = 5002,
  INVALID_NAME = 5003,
  MAXIMUM_FILE = 5004,
  FULL_DISK = 5005,
  PERMISSION_ERROR = 5006,
}

export enum SERVER_EXIT_CODE {
  INTERNAL_SERVER_ERROR = 6001,
}

export enum UNKNOW_EXIT_CODE {
  UNKNOW_ERROR = 9001,
}